+++ apps/dashboard/src/components/LeftPanel/ElementsSection.tsx
@@ -1,0 +1,53 @@
+import type { DragEndEvent } from "@dnd-kit/core";
+import { DndContext, KeyboardSensor, PointerSensor, closestCenter, useSensor, useSensors } from "@dnd-kit/core";
+import { SortableContext, arrayMove, sortableKeyboardCoordinates, verticalListSortingStrategy } from "@dnd-kit/sortable";
+import { restrictToVerticalAxis } from "@dnd-kit/modifiers";
+import { useOg } from "../OgPlayground";
+import { ElementRow } from "./ElementRow";
+
+export function ElementsSection() {
+  const { elements, setElements } = useOg()
+  const sensors = useSensors(
+    useSensor(PointerSensor, {
+      activationConstraint: {
+        distance: 5,
+      }
+    }),
+    useSensor(KeyboardSensor, {
+      coordinateGetter: sortableKeyboardCoordinates,
+    })
+  );
+
+  function handleDragEnd(event: DragEndEvent) {
+    const { active, over } = event;
+
+    if (!over) {
+      return;
+    }
+
+    if (active.id !== over.id) {
+      const oldIndex = elements.findIndex(element => element.id === active.id)
+      const newIndex = elements.findIndex(element => element.id === over.id)
+
+      const newElements = arrayMove(elements, oldIndex, newIndex);
+      setElements(newElements);
+    }
+  }
+
+  return (
+    <>
+      <p className="text-xs text-gray-600">Elements</p>
+      <div className="flex flex-col-reverse w-full">
+        <DndContext collisionDetection={closestCenter} modifiers={[restrictToVerticalAxis]} onDragEnd={handleDragEnd} onDragOver={handleDragEnd} sensors={sensors}>
+          <SortableContext items={elements} strategy={verticalListSortingStrategy}>
+            {elements.map(element => (
+              <ElementRow element={element} key={element.id} />
+            ))}
+          </SortableContext>
+        </DndContext>
+      </div>
+    </>
+  )
+}
+
+
