+++ apps/dashboard/src/lib/colors.ts
@@ -1,0 +1,10 @@
+export function hexToRgba(hex: string, alpha: number) {
+  const bigint = parseInt(hex.replace('#', ''), 16);
+  const r = (bigint >> 16) & 255;
+  const g = (bigint >> 8) & 255;
+  const b = bigint & 255;
+
+  return `rgba(${r}, ${g}, ${b}, ${alpha / 100})`;
+}
+
+
